+++ peerprep-fe/src/types/types.ts
@@ -17,4 +17,18 @@ interface ProblemDialogData {
   description: string;
 }
 
-export type { Problem, ProblemDialogData };
+interface FilterBadgeProps {
+  filterType: "difficulty" | "status" | "topics";
+  value: string;
+  onRemove: (filterType: string, value: string) => void;
+}
+
+interface FilterSelectProps {
+  placeholder: string;
+  options: { value: string; label: string }[];
+  onChange: (value: string) => void;
+  value: string | string[];
+  isMulti?: boolean;
+}
+
+export type { Problem, ProblemDialogData, FilterBadgeProps, FilterSelectProps };
